Refactor file-checking function to Actions class
Moved the `check_current_files` function to the Actions class and updated references to this method across multiple modules. This change improves code modularity by consolidating the file-checking functionality into a single class.
